+++ packages/icons/js/library-ui/icons/index.js
@@ -95,6 +95,7 @@ export { default as BlockSocialLinkIcon } from './block-social-link-icon.svg';
 export { default as BlockSocialLinkName } from './block-social-link-name.svg';
 export { default as Bolb } from './bolb.svg';
 export { default as BlockTypes } from './block-types.svg';
+export { default as BlockVideoCaption } from './block-video-caption.svg';
 export { default as BorderRadius } from './border-radius.svg';
 export { default as BreadcrumbSeparator } from './breadcrumb-separator.svg';
 export { default as BreadcrumbText } from './breadcrumb-text.svg';